+++ android/src/main/kotlin/com/eopeter/fluttermapboxnavigation/models/MapBoxLocation.kt
@@ -1,4 +1,4 @@
-package com.eopeter.flutter_mapbox_navigation.models
+package com.eopeter.fluttermapboxnavigation.models
 
 class MapBoxLocation(val name: String = "", val latitude: Double?, val longitude: Double?) {
