+++ android/src/main/kotlin/com/eopeter/fluttermapboxnavigation/models/WaypointSet.kt
@@ -1,4 +1,4 @@
-package com.eopeter.flutter_mapbox_navigation.models
+package com.eopeter.fluttermapboxnavigation.models
 
 import com.mapbox.api.directions.v5.models.RouteOptions
 import com.mapbox.geojson.Point
