+++ example/android/app/build.gradle
@@ -45,7 +45,7 @@ android {
 
     defaultConfig {
         // TODO: Specify your own unique Application ID (https://developer.android.com/studio/build/application-id.html).
-        applicationId "com.eopeter.flutter_mapbox_navigation_example"
+        applicationId "com.eopeter.fluttermapboxnavigationexample"
         // You can update the following values to match your application needs.
         // For more information, see: https://docs.flutter.dev/deployment/android#reviewing-the-gradle-build-configuration.
         minSdkVersion 21
@@ -64,7 +64,7 @@ android {
     lint {
         disable 'InvalidPackage'
     }
-    namespace 'com.eopeter.flutter_mapbox_navigation_example'
+    namespace 'com.eopeter.fluttermapboxnavigationexample'
 }
 
 flutter {
